[server] add authz to TaskService (#1432)
* [server] add authz to TaskService

* [spi] remove legacy methods in TaskManager

* [spi] simplify TaskInfo#createTaskDto method

* add namespace to created tasks

* update architecture tests

+++ thirdeye-server/src/main/java/ai/startree/thirdeye/service/TaskService.java
@@ -46,7 +46,7 @@ public class TaskService extends CrudService<TaskApi, TaskDTO> {
     this.taskManager = taskManager;
   }
 
-  // Operation not supported to prevent create of tasks
+  // operation not supported to prevent creation of tasks via API
   @Override
   protected void prepareCreatedDto(final ThirdEyeServerPrincipal principal, final TaskDTO taskDTO) {
     throw new UnsupportedOperationException();
@@ -58,7 +58,7 @@ public class TaskService extends CrudService<TaskApi, TaskDTO> {
   }
 
   public void purge(final ThirdEyeServerPrincipal principal, final int nDaysToDelete, final int limit) {
-    // fixme cyril add authz
+    authorizationManager.ensureHasRootAccess(principal);
     taskManager.purge(Duration.ofDays(nDaysToDelete), limit);
   }
 }
